% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score.R
\name{calc_risk}
\alias{calc_risk}
\title{Compute a Risk Score for a Company}
\usage{
calc_risk(bname)
}
\arguments{
\item{bname}{The name of a business.}
}
\value{

}
\description{
Compute a Risk Score for a Company
}
\examples{

bname <- "MML Properties Ltd"
profile_risk(bname)
}
