% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{closest_business_name}
\alias{closest_business_name}
\title{Find the closest matching Business}
\usage{
closest_business_name(input_value)
}
\arguments{
\item{input_value}{string pattern to search business name}
}
\value{
Closest matching business name
}
\description{
Find the closest matching Business
}
\examples{
input_value = "Gyoza Bar"
closest_business_name(input_value)

}
