% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{data_statscan}
\alias{data_statscan}
\title{Extract Network Data for Statstcan Inter-Corporate Ownership}
\usage{
data_statscan(
  path = "data-processed/combined_data.csv",
  path2 = "data-raw/linked_corp_data.csv"
)
}
\arguments{
\item{path}{The path to the linked data.}

\item{path2}{The path to the original statscan data.}
}
\value{
A list containing an edge list and a nodelist.
}
\description{
Extract Network Data for Statstcan Inter-Corporate Ownership
}
