% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{data_vbr}
\alias{data_vbr}
\title{Read-in the City-of-Vancouver Business License Data and Save to Package}
\usage{
data_vbr(path = "data-raw/business-licences.csv", save_to_pkg = TRUE)
}
\arguments{
\item{path}{A path to a csv file}

\item{save_to_pkg}{Should the data be saved as an Rda file in the data
directory (as 'vbr')?}
}
\value{
The cleaned business registry data. Saved as \code{vbr} if save_to_pkg =
TRUE.
}
\description{
This package takes existing CSV files and saves the as .Rda after data
cleaning.
}
\examples{
\dontrun{
data_vbr(save_to_pkg = FALSE)
}
}
