% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_score.R
\name{.profile_risk}
\alias{.profile_risk}
\title{Compute a Risk Score for a Company}
\usage{
.profile_risk(bname)
}
\arguments{
\item{bname}{The name of a business.}
}
\value{

}
\description{
Compute a Risk Score for a Company
}
