% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_graph.R
\name{get_ego}
\alias{get_ego}
\title{Create an Ego Network Given a Query Point}
\usage{
get_ego(bname, qp = NULL, order = 20)
}
\arguments{
\item{bname}{The name of the business to search}

\item{qp}{The query point (qp), expressed as an ID}

\item{order}{the size of the ego network. Defaults to 20.}
}
\value{

}
\description{
Create an Ego Network Given a Query Point
}
\examples{
network <- get_ego(qp = "B7")

bname <- "MML Properties Ltd"
get_ego(bname)
}
