% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_history_comment}
\alias{get_history_comment}
\title{Fetches comment about history}
\usage{
get_history_comment(bname)
}
\arguments{
\item{bname}{str indicates business name}
}
\value{

}
\description{
Fetches comment about history
}
\examples{
bname = "Gyoza Bar Ltd"
get_history_comment(bname)

}
