% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_location_comment}
\alias{get_location_comment}
\title{Fetches comment about location}
\usage{
get_location_comment(bname)
}
\arguments{
\item{bname}{str indicates business name}
}
\value{

}
\description{
Fetches comment about location
}
\examples{
bname = "Gyoza Bar Ltd"
get_location_comment(bname)

}
