% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_missing_data_comment}
\alias{get_missing_data_comment}
\title{Fetches comment about missing data}
\usage{
get_missing_data_comment(bname)
}
\arguments{
\item{bname}{str indicates business name}
}
\value{

}
\description{
Fetches comment about missing data
}
\examples{
bname = "Gyoza Bar Ltd"
get_missing_data_comment(bname)

}
