% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_operative_comment}
\alias{get_operative_comment}
\title{Fetches comment about operation}
\usage{
get_operative_comment(bname)
}
\arguments{
\item{bname}{str indicates business name}
}
\value{

}
\description{
Fetches comment about operation
}
\examples{
bname = "Gyoza Bar Ltd"
get_operative_comment(bname)

}
