% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{get_overall_risk_score_card}
\alias{get_overall_risk_score_card}
\title{Calculates risk score}
\usage{
get_overall_risk_score_card(bname)
}
\arguments{
\item{bname}{str indicates business name}
}
\value{

}
\description{
Calculates risk score
}
\examples{
bname = "Gyoza Bar Ltd"
get_overall_risk_score_card(bname)

}
