% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{make_co_type}
\alias{make_co_type}
\title{Qualifies most appropriate business type of the company}
\usage{
make_co_type(input_value)
}
\arguments{
\item{input_value}{str indicating Business Name}
}
\value{

}
\description{
Qualifies most appropriate business type of the company
}
\examples{
input_value="Gyoza Bar Ltd"
make_co_type(input_value)

}
