% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{make_related_co_table}
\alias{make_related_co_table}
\title{Generate a table of related company from stats Canada}
\usage{
make_related_co_table(input_value)
}
\arguments{
\item{input_value}{str indicating Business Name}
}
\value{

}
\description{
Generate a table of related company from stats Canada
}
\examples{
input_value="Gyoza Bar Ltd"
make_related_co_table(input_value)

}
