% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{make_score_card}
\alias{make_score_card}
\title{Title Calculates risk score}
\usage{
make_score_card(bname)
}
\arguments{
\item{bname}{str indicates business name}
}
\value{

}
\description{
Title Calculates risk score
}
\examples{
bname = "Gyoza Bar Ltd"
make_score_card(bname)

}
