% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{num_emp_plot}
\alias{num_emp_plot}
\title{Title Plotting the number of employees per year}
\usage{
num_emp_plot(input_value)
}
\arguments{
\item{input_value}{str indicating Business Name}
}
\value{

}
\description{
Title Plotting the number of employees per year
}
\examples{
input_value="Gyoza Bar Ltd"
num_emp_plot(input_value)

}
