% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_graph.R
\name{viz}
\alias{viz}
\title{Plot a Graph Using VisNetwork}
\usage{
viz(network)
}
\arguments{
\item{network}{The igraph object}
}
\value{

}
\description{
Plot a Graph Using VisNetwork
}
\examples{
network <- filter(g, BusinessName == "Gyoza Bar Ltd")
viz(network)
}
