% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_graph.R
\name{viz_graph}
\alias{viz_graph}
\title{Plot a Temporal Ego Network of a Business' Licenses}
\usage{
viz_graph(
  bname = "Gyoza Bar",
  regex = TRUE,
  ignore_case = TRUE,
  rule_filters = character(0),
  max_size = 2500
)
}
\arguments{
\item{bname}{The business name to search}
}
\value{

}
\description{
Plot a Temporal Ego Network of a Business' Licenses
}
