% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{build_graph}
\alias{build_graph}
\title{Build the Business Graph}
\usage{
build_graph(vbr = data_vbr(), statscan = data_statscan(), save_to_pkg = TRUE)
}
\arguments{
\item{save_to_pkg}{Whether or not the graph should be saved to the package
data directory. Defaults to TRUE.}

\item{vrb}{The Vancouver business registry dataset (saved in the package as
\code{vbr})}
}
\value{

}
\description{
Build the Business Graph
}
\examples{
\dontrun{
build_graph(save_to_pkg = FALSE)
}
}
