% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_graph.R
\name{viz_graph}
\alias{viz_graph}
\title{Plot a Temporal Ego Network of a Business' Licenses}
\usage{
viz_graph(bname = "Gyoza Bar Ltd")
}
\arguments{
\item{bname}{The business name to search}
}
\value{

}
\description{
Plot a Temporal Ego Network of a Business' Licenses
}
\examples{
bname <- "MML Properties Ltd"
viz_graph(bname)
}
